+++ tests/Period/IntervalRelationTest.php
@@ -992,4 +992,64 @@ class IntervalRelationTest extends TestCase
             }
         }
     }
+
+    public function testSubstractWithOverlappingUnequalPeriods(): void
+    {
+        $periodA = Period::after('2000-01-01 10:00:00', '8 HOURS');
+        $periodB = Period::after('2000-01-01 14:00:00', '6 HOURS');
+
+        $diff1 = $periodA->substract($periodB);
+
+        self::assertCount(1, $diff1);
+        self::assertEquals($periodA->getStartDate(), $diff1[0]->getStartDate());
+        self::assertEquals($periodB->getStartDate(), $diff1[0]->getEndDate());
+
+        $diff2 = $periodB->substract($periodA);
+
+        self::assertCount(1, $diff2);
+        self::assertEquals($periodA->getEndDate(), $diff2[0]->getStartDate());
+        self::assertEquals($periodB->getEndDate(), $diff2[0]->getEndDate());
+    }
+
+    public function testSubstractWithSeparatePeriods(): void
+    {
+        $periodA = Period::after('2000-01-01 10:00:00', '4 HOURS');
+        $periodB = Period::after('2000-01-01 15:00:00', '3 HOURS');
+
+        $diff1 = $periodA->substract($periodB);
+
+        self::assertCount(1, $diff1);
+        self::assertTrue($diff1[0]->equals($periodA));
+
+        $diff2 = $periodB->substract($periodA);
+
+        self::assertCount(1, $diff2);
+        self::assertTrue($diff2[0]->equals($periodB));
+    }
+
+    public function testSubstractWithOnePeriodContainedInAnother(): void
+    {
+        $periodA = Period::after('2000-01-01 10:00:00', '8 HOURS');
+        $periodB = Period::after('2000-01-01 15:00:00', '1 HOUR');
+
+        $diff1 = $periodA->substract($periodB);
+
+        self::assertCount(2, $diff1);
+        self::assertEquals($periodA->getStartDate(), $diff1[0]->getStartDate());
+        self::assertEquals($periodB->getStartDate(), $diff1[0]->getEndDate());
+        self::assertEquals($periodB->getEndDate(), $diff1[1]->getStartDate());
+        self::assertEquals($periodA->getEndDate(), $diff1[1]->getEndDate());
+
+        $diff2 = $periodB->substract($periodA);
+
+        self::assertCount(0, $diff2);
+    }
+
+    public function testSubstractWithEqualPeriodObjec(): void
+    {
+        $periodA = Period::after('2000-01-01 10:00:00', '8 HOURS');
+        $diff = $periodA->substract($periodA);
+
+        self::assertCount(0, $diff);
+    }
 }